Rename ospd to osp, minor refactoring for clarity
