--- rust/examples/if-assignment.nasl
@@ -1,2 +1,0 @@
-if (a = 1) display (a);
-if ((a = 1)) display('a', a);
